rename autoconfig to uc_check
